🔨 Refactor: navbar 사이드바 이슈 반영
